paiagram_core

Module entry

Structs

AdjustEntryMode

Changes the entry’s mode This would trigger a schedule estimate recalculation

ChangeEntryStop

Changes the entry’s stop. This would trigger a route recalculation

EntryBundle

Bundle for easy spawning

EntryEstimate

The estimated arrival and departure times of the entry. This is not a hard requirement for entries.

EntryMode

The entry’s TravelModes.

EntryPlugin

EntryQuery

EntryQueryItem

Automatically generated [WorldQuery](:: bevy :: ecs::query::WorldQuery) item type for EntryQuery, returned when iterating over query results.

EntryStop

Where the vehicle stops. The stop could be a station, or a platform that belongs to the station.

IsDerivedEntry

Marker component given to derived entries

Enums

EntryModeAdjustment

TravelMode

Travel mode for entries