paiagram_core

Module graph

Definitions for the graph.

Modules

arrange

Structs

AddIntervalPair

Graph

The graph. Graph stores the node using an entity and the edge also using an entity entity in a DiGraphMap using [EntityHash].

GraphIntervalSpatialIndex

GraphIntervalSpatialIndexState 🔒

GraphIntervalSpatialSample

GraphPlugin

The graph plugin. Remember to register this plugin.

GraphSpatialIndex

GraphSpatialIndexState 🔒

IntervalSpatialIndexedEntity 🔒

Node

NodeCoor

The position of the node.

SpatialIndexedEntity 🔒

Constants

EARTH_RADIUS_METERS 🔒

WEB_MERCATOR_MAX_LAT 🔒

Functions

add_interval_pair 🔒

apply_graph_interval_spatial_index_task 🔒

apply_graph_spatial_index_task 🔒

check_stations_in_graph 🔒

debug_graph_set_diff 🔒

lon_lat_to_xy

mark_graph_interval_spatial_index_dirty 🔒

mark_graph_spatial_index_dirty 🔒

merge_station_by_name

start_graph_interval_spatial_index_rebuild 🔒

start_graph_spatial_index_rebuild 🔒

update_graph_on_interval_removal 🔒

update_graph_on_station_removal 🔒

xy_to_lon_lat