paiagram_core

Module station

Structs

IsDepot

Marker component for depots.

IsExternalStation

Marker component for stations that are not managed by the current diagram. Note that being “external” doesn’t mean that it can be excluded from crate::graph::Graph.

ParentStationOrStation

ParentStationOrStationItem

Automatically generated [WorldQuery](:: bevy :: ecs::query::WorldQuery) item type for ParentStationOrStation, returned when iterating over query results.

Platform

PlatformEntries

Entries that passes this platform.

PlatformQuery

PlatformQueryItem

Automatically generated [WorldQuery](:: bevy :: ecs::query::WorldQuery) item type for PlatformQuery, returned when iterating over query results.

Station

A station in the network. A station would also host a default platform.

StationBundle

Spawn with [`with_children()``]

StationPlugin

StationQuery

StationQueryItem

Automatically generated [WorldQuery](:: bevy :: ecs::query::WorldQuery) item type for StationQuery, returned when iterating over query results.

Type Aliases

Platforms

A type alias to make things clearer.

StationEntries

Entries that visit this station. This does not contain all entries that visit this station, since a station may contain extra platforms, and each platform would also record a set of entries that visit the platform, not the station. See PlatformEntries for further details.